refactor: extract tls setup into http mod & refactor remaining mods
